Massive changes and improvements
Including, but not limited to: Context menu for Cue list,
working CueInfor views for both the active and selected cue, and Add Cue Button, a working cue editor with proper saving, refactored MainWindow code.

Still to do for these features is the insertion of new cues based on their alhpa numeric codes.